refactor [#237] 애플/카카오 로그인 프로세스 불일치 및 추상화를 위해 인터페이스 수정 및 레지스트리 작성, domain 하위로 폴더 이동
