+++ src/main/java/org/sopt/confeti/api/auth/controller/AuthController.java
@@ -4,11 +4,9 @@ import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
 import org.sopt.confeti.api.auth.dto.LoginRequest;
 import org.sopt.confeti.api.auth.facade.AuthFacade;
-import org.sopt.confeti.auth.Token;
-import org.sopt.confeti.auth.command.LoginCommand;
-import org.sopt.confeti.auth.dto.LoginResult;
-import org.sopt.confeti.domain.user.constant.Role;
-import org.sopt.confeti.global.annotation.Permission;
+import org.sopt.confeti.domain.auth.Token;
+import org.sopt.confeti.domain.auth.command.LoginCommand;
+import org.sopt.confeti.domain.auth.dto.LoginResult;
 import org.sopt.confeti.global.annotation.UserId;
 import org.sopt.confeti.global.common.BaseResponse;
 import org.sopt.confeti.global.message.SuccessMessage;
@@ -24,16 +22,15 @@ public class AuthController {
 
     private final AuthFacade authFacade;
 
-    @PostMapping("/login")
+    @PostMapping("/kakao/login")
     public ResponseEntity<BaseResponse<?>> login(
            @Valid @RequestBody LoginRequest request
     ) {
-        LoginResult result = authFacade.login(LoginCommand.from(request));
+        LoginResult result = authFacade.login(new LoginCommand(request.provider(), request.redirectUrl(), request.code()));
         return ApiResponseUtil.success(SuccessMessage.SUCCESS, result);
     }
 
-    @Permission(role =  {Role.GENERAL})
-    @PostMapping("/reissue")
+    @PostMapping("/kakao/reissue")
     public ResponseEntity<BaseResponse<?>> reissue(
             @RequestHeader(HttpHeaders.AUTHORIZATION) String refreshToken
     ) {
@@ -41,8 +38,7 @@ public class AuthController {
         return ApiResponseUtil.success(SuccessMessage.SUCCESS, token);
     }
 
-    @Permission(role = {Role.GENERAL})
-    @PostMapping("/logout")
+    @PostMapping("/kakao/logout")
     public ResponseEntity<BaseResponse<?>> logout(@UserId Long userId){
         authFacade.logout(userId);
         return ApiResponseUtil.success(SuccessMessage.SUCCESS);
