+++ src/main/java/org/sopt/confeti/api/auth/facade/AuthFacade.java
@@ -1,12 +1,12 @@
 package org.sopt.confeti.api.auth.facade;
 
 import lombok.RequiredArgsConstructor;
-import org.sopt.confeti.auth.LoginService;
-import org.sopt.confeti.auth.LogoutService;
-import org.sopt.confeti.auth.ReissueService;
-import org.sopt.confeti.auth.Token;
-import org.sopt.confeti.auth.command.LoginCommand;
-import org.sopt.confeti.auth.dto.LoginResult;
+import org.sopt.confeti.domain.auth.LoginService;
+import org.sopt.confeti.domain.auth.LogoutService;
+import org.sopt.confeti.domain.auth.ReissueService;
+import org.sopt.confeti.domain.auth.Token;
+import org.sopt.confeti.domain.auth.command.LoginCommand;
+import org.sopt.confeti.domain.auth.dto.LoginResult;
 import org.sopt.confeti.global.annotation.Facade;
 import org.springframework.transaction.annotation.Transactional;
