--- src/main/java/org/sopt/confeti/auth/Token.java
@@ -1,7 +1,0 @@
-package org.sopt.confeti.auth;
-
-public record Token (
-        String accessToken,
-        String refreshToken
-){
-}
