--- src/main/java/org/sopt/confeti/auth/dto/apple/ApplePublicKey.java
@@ -1,11 +1,0 @@
-package org.sopt.confeti.auth.dto.apple;
-
-public record ApplePublicKey(
-        String kty,
-        String kid,
-        String use,
-        String alg,
-        String n,
-        String e
-) {
-}
