--- src/main/java/org/sopt/confeti/auth/dto/apple/ApplePublicKeys.java
@@ -1,8 +1,0 @@
-package org.sopt.confeti.auth.dto.apple;
-
-import java.util.List;
-
-public record ApplePublicKeys(
-        List<ApplePublicKey> keys
-) {
-}
