--- src/main/java/org/sopt/confeti/auth/dto/apple/AppleTokenRequestParams.java
@@ -1,12 +1,0 @@
-package org.sopt.confeti.auth.dto.apple;
-
-public record AppleTokenRequestParams(
-        String clientId,
-        String clientSecret,
-        String grantType,
-        String code
-) {
-    public static AppleTokenRequestParams of(String clientId, String clientSecret, String grantType, String code) {
-        return new AppleTokenRequestParams(clientId, clientSecret, grantType, code);
-    }
-}
