--- src/main/java/org/sopt/confeti/auth/dto/kakao/KakaoAccount.java
@@ -1,6 +1,0 @@
-package org.sopt.confeti.auth.dto.kakao;
-
-public record KakaoAccount (
-        KakaoProfile profile
-){
-}
