--- src/main/java/org/sopt/confeti/auth/jwt/JwtProperties.java
@@ -1,11 +1,0 @@
-package org.sopt.confeti.auth.jwt;
-
-import org.springframework.boot.context.properties.ConfigurationProperties;
-
-@ConfigurationProperties(prefix = "jwt")
-public record JwtProperties (
-        String secretKey,
-        int accessTokenValidTime,
-        int refreshTokenValidTime
-){
-}
