--- src/main/java/org/sopt/confeti/auth/jwt/JwtPropertiesConfiguration.java
@@ -1,9 +1,0 @@
-package org.sopt.confeti.auth.jwt;
-
-import org.springframework.boot.context.properties.ConfigurationPropertiesScan;
-import org.springframework.context.annotation.Configuration;
-
-@Configuration
-@ConfigurationPropertiesScan(basePackages = "org.sopt.confeti.auth.jwt")
-public class JwtPropertiesConfiguration {
-}
