--- src/main/java/org/sopt/confeti/auth/jwt/KeyGenerator.java
@@ -1,16 +1,0 @@
-package org.sopt.confeti.auth.jwt;
-
-import io.jsonwebtoken.io.Decoders;
-import javax.crypto.SecretKey;
-import org.springframework.stereotype.Component;
-
-import javax.crypto.spec.SecretKeySpec;
-
-@Component
-public class KeyGenerator {
-    private static final String HMACSHA256 = "HmacSHA256";
-    public SecretKey getKeyFromString(String keyString) {
-        byte[] keyBytes = Decoders.BASE64.decode(keyString);
-        return new SecretKeySpec(keyBytes, HMACSHA256);
-    }
-}
