+++ src/main/java/org/sopt/confeti/domain/auth/Token.java
@@ -1,0 +1,7 @@
+package org.sopt.confeti.domain.auth;
+
+public record Token (
+        String accessToken,
+        String refreshToken
+){
+}
