+++ src/main/java/org/sopt/confeti/domain/auth/dto/KakaoAccount.java
@@ -1,0 +1,6 @@
+package org.sopt.confeti.domain.auth.dto;
+
+public record KakaoAccount (
+        KakaoProfile profile
+){
+}
