+++ src/main/java/org/sopt/confeti/domain/auth/dto/KakaoLoginParams.java
@@ -1,0 +1,12 @@
+package org.sopt.confeti.domain.auth.dto;
+
+import org.sopt.confeti.domain.auth.command.LoginCommand;
+
+public record KakaoLoginParams(
+        String redirectUrl,
+        String code
+){
+    public static KakaoLoginParams from(LoginCommand command) {
+        return new KakaoLoginParams(command.redirectUrl(), command.code());
+    }
+}
