+++ src/main/java/org/sopt/confeti/domain/auth/dto/kakao/KakaoProfile.java
@@ -1,0 +1,13 @@
+package org.sopt.confeti.domain.auth.dto;
+
+import com.fasterxml.jackson.annotation.JsonProperty;
+
+public record KakaoProfile(
+
+        @JsonProperty("nickname")
+        String nickname,
+
+        @JsonProperty("profile_image_url")
+        String profileImageUrl
+) {
+}
