+++ src/main/java/org/sopt/confeti/domain/auth/dto/kakao/KakaoSocialInfoResult.java
@@ -1,0 +1,10 @@
+package org.sopt.confeti.domain.auth.dto;
+
+import com.fasterxml.jackson.annotation.JsonProperty;
+
+public record KakaoSocialInfoResult(
+        String id,
+        @JsonProperty("kakao_account")
+        KakaoAccount kakaoAccount
+){
+}
