+++ src/main/java/org/sopt/confeti/domain/auth/jwt/JwtProperties.java
@@ -1,0 +1,11 @@
+package org.sopt.confeti.domain.auth.jwt;
+
+import org.springframework.boot.context.properties.ConfigurationProperties;
+
+@ConfigurationProperties(prefix = "jwt")
+public record JwtProperties (
+        String secretKey,
+        int accessTokenValidTime,
+        int refreshTokenValidTime
+){
+}
