+++ src/main/java/org/sopt/confeti/domain/auth/jwt/JwtPropertiesConfiguration.java
@@ -1,0 +1,9 @@
+package org.sopt.confeti.domain.auth.jwt;
+
+import org.springframework.boot.context.properties.ConfigurationPropertiesScan;
+import org.springframework.context.annotation.Configuration;
+
+@Configuration
+@ConfigurationPropertiesScan(basePackages = "org.sopt.confeti.auth.jwt")
+public class JwtPropertiesConfiguration {
+}
