+++ src/main/java/org/sopt/confeti/domain/auth/jwt/JwtTokenExtractor.java
@@ -1,0 +1,36 @@
+package org.sopt.confeti.domain.auth.jwt;
+
+import io.jsonwebtoken.Jwts;
+import lombok.RequiredArgsConstructor;
+import org.sopt.confeti.domain.user.constant.Role;
+import org.springframework.stereotype.Component;
+
+@Component
+@RequiredArgsConstructor
+public class JwtTokenExtractor {
+
+    private static final String JWT_CLAIM_ROLE = "role";
+
+    private final JwtProperties jwtProperties;
+    private final KeyGenerator keyGenerator;
+
+    public String getSubject(String token) {
+        return Jwts.parserBuilder()
+                .setSigningKey(keyGenerator.getKeyFromString(jwtProperties.secretKey()))
+                .build()
+                .parseClaimsJws(token)
+                .getBody()
+                .getSubject();
+    }
+
+    public Role getRole(String token) {
+        String role = Jwts.parserBuilder()
+                .setSigningKey(keyGenerator.getKeyFromString(jwtProperties.secretKey()))
+                .build()
+                .parseClaimsJws(token)
+                .getBody()
+                .get(JWT_CLAIM_ROLE, String.class);
+
+        return Role.from(role);
+    }
+}
