+++ src/main/java/org/sopt/confeti/global/annotation/interceptor/PermissionInterceptor.java
@@ -6,8 +6,8 @@ import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 import java.util.Set;
 import lombok.RequiredArgsConstructor;
-import org.sopt.confeti.auth.jwt.JwtTokenExtractor;
-import org.sopt.confeti.auth.jwt.TokenParser;
+import org.sopt.confeti.domain.auth.jwt.JwtTokenExtractor;
+import org.sopt.confeti.domain.auth.jwt.TokenParser;
 import org.sopt.confeti.domain.user.constant.Role;
 import org.sopt.confeti.global.annotation.Interceptor;
 import org.sopt.confeti.global.annotation.Permission;
