+++ src/main/java/org/sopt/confeti/global/resolver/user/UserIdArgumentResolver.java
@@ -4,8 +4,8 @@ import io.jsonwebtoken.ExpiredJwtException;
 import io.jsonwebtoken.JwtException;
 import jakarta.servlet.http.HttpServletRequest;
 import lombok.RequiredArgsConstructor;
-import org.sopt.confeti.auth.jwt.JwtTokenExtractor;
-import org.sopt.confeti.auth.jwt.TokenParser;
+import org.sopt.confeti.domain.auth.jwt.JwtTokenExtractor;
+import org.sopt.confeti.domain.auth.jwt.TokenParser;
 import org.sopt.confeti.global.annotation.UserId;
 import org.sopt.confeti.global.exception.UnauthorizedException;
 import org.springframework.http.HttpHeaders;
